"Enhanced graph visualization and data handling"
In this commit, we've refined the graph visualization in both 'qt_dash.py' and 'dropdowns.py'. This includes subtle aesthetic upgrades to highlight minor grid lines on the X and Y axes. Additionally, we have refined the handling of scatter trace addition to include an optional 'secondary_y' parameter to show boolean data