Handle multiple files when loading data
When multiple RTI files was read, we previously only read the first database and looked into the rti_json_data and populated the columns and channels from one file.

This caused problems if first file did not contain any data.

Now all files is being read to populate the groups and channels
.