Ensure data presence before making plotly figure
This commit adds conditions to handle the scenario of absent timeline data in novos_processes.py and fig2 data in dropdowns.py. It helps to avoid potential exceptions caused by unexpected access to missing data, improving the robustness of the data plotting and present fallback messages when the data is not available.