Add custom color option to plot lines
Added an option to allow users to select custom line colors for the plots in the application. Changes were applied to the qt_dash.py, my_custom_classes.py, and dropdowns.py files. By enabling the color customization, it provides a better user experience and makes the plots easier to differentiate when multiple lines are present.